+++ docs/source/user_guide/changelog.rst
@@ -24,6 +24,11 @@ Changelog
 
 Authors since :ref:`5.0.0`: Jordi Torrents (jordi.torrentsm@gmail.com)
 
+Unreleased
+==========
+
+- Remove parameter ``IDENTITY_TRANSFER`` setting it to ``True`` if a knowledge transfer folder is provided.
+
 6.0.12
 ======